dom/l10n/tests/mochitest/dom_localization/test_overlay_missing_all.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test DOMLocalization's DOMOverlay functionality</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript">
  "use strict";
  async function* generateBundles(resourceIds) {
    const bundle = new FluentBundle("en-US");
    // No translations!
    yield bundle;
  }

  SimpleTest.waitForExplicitFinish();
  addLoadEvent(async () => {
    const domLoc = new DOMLocalization(
      [],
      false,
      { generateBundles },
    );

    await domLoc.translateFragment(document.body).then(() => {
      ok(false, "Expected translateFragment to throw on missing l10n-id");
    }, () => {
      ok(true, "Expected translateFragment to throw on missing l10n-id");
    });
    SimpleTest.finish();
  });
  </script>
</head>
<body>
  <p data-l10n-id="title">
    <a href="http://www.mozilla.org"></a>
    <a href="http://www.firefox.com"></a>
    <a href="http://www.w3.org"></a>
  </p>
</body>
</html>
